// types.ts
export enum Karat {
  K18 = 18,
  K21 = 21,
  K24 = 24,
}

export enum ItemStatus {
  IN_STORE = 'IN_STORE',
  WITH_REP = 'WITH_REP',
  SOLD = 'SOLD',
}

export enum TransactionType {
  SALE = 'SALE',
  PURCHASE = 'PURCHASE',
  PAYMENT_OUT = 'PAYMENT_OUT',
  DEBT_PAYMENT = 'DEBT_PAYMENT',
}

export enum PaymentMethod {
  CASH = 'CASH',
  INSTAPAY = 'INSTAPAY',
  DEFERRED = 'DEFERRED',
  SPLIT = 'SPLIT',
}
// ...
export interface Transaction {
  id: string;
  invoiceId?: string;
  type: TransactionType;
  itemId?: string;
  itemName?: string;
  weight: number;
  qty?: number;
  karat: Karat;
  goldPricePerGram: number;
  workmanship: number;
  totalPrice: number;
  paidAmount?: number;
  cashAmount?: number;
  instapayAmount?: number;
  paymentMethod: PaymentMethod;
  customerName?: string;
  customerPhone?: string;
  customerId?: string;
  repId?: string;
  safeId?: string;
  date: number;
  notes?: string;
  createdBy: string;
  source?: 'DESKTOP' | 'MOBILE';
  updatedAt?: number;
  shop?: string;
}